# openapi/openapi.yaml
openapi: '3.1.0'
info:
  title: Narrex API
  version: '0.1.0'
  description: |
    API for Narrex — a visual novel editor with AI scene generation.

    Phase 1 covers the core loop: project creation with auto-structuring,
    multi-track timeline, character map, and scene-level AI draft generation.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/auth/google:
    get:
      operationId: initiateGoogleAuth
      tags: [auth]
      summary: Redirect to Google OAuth consent screen
      security: []
      parameters:
        - name: redirect_uri
          in: query
          required: false
          description: URL to redirect after authentication completes
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to Google OAuth
          headers:
            Location:
              schema:
                type: string
              description: Google OAuth consent URL
  /v1/auth/google/callback:
    get:
      operationId: handleGoogleCallback
      tags: [auth]
      summary: Handle Google OAuth callback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to web app with tokens set
          headers:
            Location:
              schema:
                type: string
              description: Web app URL
            Set-Cookie:
              schema:
                type: string
              description: httpOnly refresh token cookie (30 days)
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      description: Uses httpOnly refresh token cookie. No request body needed.
      security: []
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfile'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Delete current user's account and all data
      description: |
        Permanently deletes the authenticated user's account and all associated
        data (projects, scenes, characters, drafts, generation logs).
        Clears the refresh token cookie. Irreversible.
      responses:
        '204':
          description: Account deleted
          headers:
            Set-Cookie:
              description: Clears the refresh token cookie
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/auth/me/avatar:
    post:
      operationId: uploadAvatar
      tags: [auth]
      summary: Upload profile avatar image
      description: |
        Accepts a multipart file upload (field name: `avatar`).
        Supported formats: JPEG, PNG, WebP. Max size: 2MB.
        The image is stored in R2 and the user's profileImageUrl is updated.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, or WebP, max 2MB)
      responses:
        '200':
          description: Updated user with new avatar URL
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/me/costs:
    get:
      operationId: getUserCosts
      tags: [generation]
      summary: Get AI generation cost summary for current user
      responses:
        '200':
          description: User cost summary
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/CostSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---- Projects -----------------------------------------------------------
  /v1/projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's projects
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Project list sorted by updatedAt descending
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createProject
      tags: [projects]
      summary: Create project with AI auto-structuring
      description: |
        Streams structuring progress via SSE. Accepts text input (JSON) or file
        upload (multipart).

        SSE event types:
        - `progress` — `{"message": "Analyzing input..."}` — status updates
        - `clarification` — `{"questions": [...]}` — input too vague; stream
          ends. Resubmit with `clarificationAnswers`.
        - `completed` — `{"data": {...workspace...}}` — project created with
          full workspace data (same shape as Workspace schema)
        - `error` — `{"message": "..."}` — structuring failed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectText'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateProjectFile'
      responses:
        '200':
          description: SSE stream of structuring progress
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project config
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project config
      description: |
        Updates story configuration fields. Scenes with existing AI-generated or
        edited drafts are automatically marked `needs_revision` when config
        changes affect generation context (genre, tone, pov, etc.).
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProject'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Soft-delete a project
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '204':
          description: Project deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/projects/{projectId}/workspace:
    get:
      operationId: getWorkspace
      tags: [projects]
      summary: Get full workspace data
      description: |
        Returns project config, all tracks, scenes (with character assignments
        but without draft content), characters, relationships, and connections
        in a single response. Use getScene for individual scene draft content.
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '200':
          description: Full workspace
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/projects/{projectId}/costs:
    get:
      operationId: getProjectCosts
      tags: [generation]
      summary: Get AI generation cost summary for a project
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '200':
          description: Project cost summary
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/CostSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Tracks -------------------------------------------------------------
  /v1/projects/{projectId}/tracks:
    post:
      operationId: createTrack
      tags: [tracks]
      summary: Create a timeline track
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrack'
      responses:
        '201':
          description: Track created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Track'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/projects/{projectId}/tracks/{trackId}:
    patch:
      operationId: updateTrack
      tags: [tracks]
      summary: Update track label or position
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/TrackIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrack'
      responses:
        '200':
          description: Track updated
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Track'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteTrack
      tags: [tracks]
      summary: Delete a track
      description: Returns 409 if track contains scenes. Delete or move scenes first.
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/TrackIdParam'
      responses:
        '204':
          description: Track deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ---- Scenes -------------------------------------------------------------
  /v1/projects/{projectId}/scenes:
    post:
      operationId: createScene
      tags: [scenes]
      summary: Create a scene on a track
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScene'
      responses:
        '201':
          description: Scene created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Scene'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/projects/{projectId}/scenes/{sceneId}:
    get:
      operationId: getScene
      tags: [scenes]
      summary: Get scene with latest draft
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/SceneIdParam'
      responses:
        '200':
          description: Scene with latest draft content
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/SceneDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateScene
      tags: [scenes]
      summary: Update scene metadata or position
      description: |
        Updates scene fields including title, plot summary, location, mood tags,
        character assignments, timeline position, and track assignment.
        Setting `trackId` moves the scene to a different track. Setting
        `characterIds` replaces all character assignments for the scene.
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/SceneIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScene'
      responses:
        '200':
          description: Scene updated
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Scene'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      operationId: deleteScene
      tags: [scenes]
      summary: Delete a scene
      description: Cascades to drafts, summary, connections, and character assignments.
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/SceneIdParam'
      responses:
        '204':
          description: Scene deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Connections --------------------------------------------------------
  /v1/projects/{projectId}/connections:
    post:
      operationId: createConnection
      tags: [connections]
      summary: Create a branch or merge connection between scenes
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnection'
      responses:
        '201':
          description: Connection created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/SceneConnection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/projects/{projectId}/connections/{connectionId}:
    delete:
      operationId: deleteConnection
      tags: [connections]
      summary: Delete a scene connection
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/ConnectionIdParam'
      responses:
        '204':
          description: Connection deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Characters ---------------------------------------------------------
  /v1/projects/{projectId}/characters:
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacter'
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /v1/projects/{projectId}/relationships:
    post:
      operationId: createRelationship
      tags: [relationships]
      summary: Create a character relationship
      description: |
        Creates a relationship between two characters. One relationship per
        character pair. Character IDs are sorted internally (a < b) to
        prevent duplicates.
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRelationship'
      responses:
        '201':
          description: Relationship created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/CharacterRelationship'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Self-issued JWT access token (15 min lifetime)

  # ---- Reusable Parameters ------------------------------------------------
  parameters:
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request syntax
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    User:
      type: object
      required: [id, email, displayName, themePreference, languagePreference, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        displayName:
          type: string
        profileImageUrl:
          type:
            - string
            - 'null'
          format: uri
        themePreference:
          type: string
          enum: [system, light, dark]
          default: system
          description: User's theme preference
        languagePreference:
          type: string
          enum: [ko, en]
          default: ko
          description: User's language preference
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    AuthTokens:
      type: object
      required: [accessToken, expiresIn]
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Token lifetime in seconds
          example: 900

    # ---- Project ----------------------------------------------------------
...
security:
  - bearerAuth: []